refactor: 工程化歌词提供商相关接口
